Zero out rewards
