+++ src/views/Stake/ClaimModal.css
@@ -1,4 +1,12 @@
 
+.Spread-capture .buy-input .Exchange-swap-section-top {
+  font-size: 16px!important;
+}
+
+.Spread-capture .buy-input {
+  border: 1px solid var(--cell-highlight);
+}
+
 .MlpSwap-content {
   display: flex;
   justify-content: space-between;
