+++ src/views/Stake/StakeV2.css
@@ -108,6 +108,10 @@
   display: none;
 }
 
+.Spread-capture .buy-input .Exchange-swap-section-top {
+  font-size: 16px!important;
+}
+
 @media (max-width: 900px) {
   .StakeV2-cards {
     grid-template-columns: 1fr;
